feat: change score values from float32 to float64 for precision
Updates all score-related fields to use float64 (double) instead of float32
to preserve precision for very small probability values (e.g., PEP values
like 1e-20 or smaller e-values).

Changes:
- score_value in additional_scores: float32 -> float64
- posterior_error_probability: float32 -> float64
- global_qvalue, pg_global_qvalue, pg_qvalue: float32 -> float64
- modification scores score_value: float32 -> float64

Updated files:
- PyArrow schema (format.py)
- Avro schemas (psm.avsc, pg.avsc, feature.avsc, peptide.avsc)
- FragPipe Score dataclass
- Documentation (format-specification.md, PSM_SCHEMA.md)

+++ qpx/core/format.py
@@ -110,7 +110,7 @@ PEPTIDE_FIELDS = [
                                                     ),
                                                     pa.field(
                                                         "score_value",
-                                                        pa.float32(),
+                                                        pa.float64(),
                                                         metadata={
                                                             "description": "Value of the score for this specific position"
                                                         },
@@ -153,7 +153,7 @@ PEPTIDE_FIELDS = [
     ),
     pa.field(
         "posterior_error_probability",
-        pa.float32(),
+        pa.float64(),
         nullable=True,
         metadata={
             "description": "Posterior error probability for the given peptide or psm match."
@@ -183,7 +183,7 @@ PEPTIDE_FIELDS = [
         pa.list_(
             pa.struct([
                 ("score_name", pa.string()),
-                ("score_value", pa.float32()),
+                ("score_value", pa.float64()),
                 ("higher_better", pa.bool_()),
             ])
         ),
@@ -359,7 +359,7 @@ FEATURE_UNIQUE_FIELDS = [
     ),
     pa.field(
         "pg_global_qvalue",
-        pa.float32(),
+        pa.float64(),
         nullable=True,
         metadata={
             "description": "Global q-value of the protein group at the experiment level"
@@ -531,7 +531,7 @@ PG_FIELDS = [
     ),
     pa.field(
         "global_qvalue",
-        pa.float32(),
+        pa.float64(),
         nullable=True,
         metadata={
             "description": "Global q-value of the protein group at the experiment level"
@@ -539,7 +539,7 @@ PG_FIELDS = [
     ),
     pa.field(
         "pg_qvalue",
-        pa.float32(),
+        pa.float64(),
         nullable=True,
         metadata={
             "description": "Protein group q-value at the run level (DIA-NN specific)"
@@ -632,7 +632,7 @@ PG_FIELDS = [
         pa.list_(
             pa.struct([
                 ("score_name", pa.string()),
-                ("score_value", pa.float32()),
+                ("score_value", pa.float64()),
                 ("higher_better", pa.bool_()),
             ])
         ),
